Extract Core.Memory from Core.Helpers
In preparation for #340.
